LayoutTests/fast/forms/ValidityState-stepMismatch-expected.txt

Check stepMismatch results for date, number, range, unsupported types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Date type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1969-12-31") is false
Normal step values
PASS stepMismatchFor("2010-02-10", "2", "2010-02-09") is true
PASS stepMismatchFor("2010-02-09", "2", "2010-02-09") is false
PASS stepMismatchFor("2010-02-11", "2", "2010-02-09") is false
PASS stepMismatchFor("1800-11-11", "3", "1800-11-09") is true
PASS stepMismatchFor("1800-11-09", "3", "1800-11-09") is false
PASS stepMismatchFor("1800-11-12", "3", "1800-11-09") is false
PASS stepMismatchFor("275760-09-13", "3", "275760-09-11") is true
PASS stepMismatchFor("275760-09-13", "2", "275760-09-11") is false
Implicit step base
PASS stepMismatchFor("1970-01-02", "2", null) is true
PASS stepMismatchFor("1970-01-03", "2", null) is false
Fractional step values
PASS stepMismatchFor("2010-02-10", "0.1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "1.1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "1.9", "2010-02-09") is true
Invalid or no step values
PASS stepMismatchFor("2010-02-10", null, "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "-1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "foo", "2010-02-09") is false
Special step value
PASS stepMismatchFor("2010-02-10", "any", "2010-02-09") is false
Disabled
PASS stepMismatchFor("2010-02-10", "2", "2010-02-09", true) is false

Number type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "1.0", "0.1") is false
Integers
PASS stepMismatchFor("1", "2", "0") is true
PASS stepMismatchFor("-3", "2", "-4") is true
PASS input.max = "5"; stepMismatchFor("5", "3", "0") is true
PASS input.value is "5"
Invalid step values
PASS stepMismatchFor("-3", "-2", "-4") is false
PASS stepMismatchFor("-3", null, "-4") is false
PASS stepMismatchFor("-3", undefined, "-4") is false
Huge numbers and small step; uncomparable
PASS stepMismatchFor("3.40282347e+38", "3", "") is false
PASS stepMismatchFor("3.40282346e+38", "3", "") is false
PASS stepMismatchFor("3.40282345e+38", "3", "") is false
Huge numbers and huge step
PASS stepMismatchFor("3.20e+38", "0.20e+38", "") is false
PASS stepMismatchFor("3.20e+38", "0.22e+38", "") is true
Fractional numbers
PASS stepMismatchFor("0.9", "0.1", "") is false
PASS stepMismatchFor("0.9", "0.1000001", "") is true
PASS stepMismatchFor("0.9", "0.1000000000000001", "") is false
PASS stepMismatchFor("1.0", "0.3333333333333333", "") is false
Rounding
PASS stepMismatchFor("5.005", "0.005", "4") is false
Disabled
PASS stepMismatchFor("1", "2", "0", true) is false

Range type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "1.0", "0.1") is false
Integers
PASS stepMismatchFor("1", "2", "0") is false
PASS stepMismatchFor("-3", "2", "-4") is false
PASS input.max = "5"; stepMismatchFor("5", "3", "0") is false
PASS input.value is "3"
Invalid step values
PASS stepMismatchFor("-3", "-2", "-4") is false
PASS stepMismatchFor("-3", null, "-4") is false
PASS stepMismatchFor("-3", undefined, "-4") is false
Huge numbers and small step; uncomparable
PASS stepMismatchFor("3.40282347e+38", "3", "") is false
PASS stepMismatchFor("3.40282346e+38", "3", "") is false
PASS stepMismatchFor("3.40282345e+38", "3", "") is false
Huge numbers and huge step
PASS stepMismatchFor("3.20e+38", "0.20e+38", "") is false
PASS stepMismatchFor("3.20e+38", "0.22e+38", "") is false
Fractional numbers
PASS stepMismatchFor("0.9", "0.1", "") is false
PASS stepMismatchFor("0.9", "0.1000001", "") is false
PASS stepMismatchFor("0.9", "0.1000000000000001", "") is false
PASS stepMismatchFor("1.0", "0.3333333333333333", "") is false
Disabled
PASS stepMismatchFor("1", "2", "0", true) is false

Unsupported types
PASS input.type = "text"; input.step = "3"; input.min = ""; input.value = "2"; input.validity.stepMismatch is false
PASS input.type = "button"; input.validity.stepMismatch is false
PASS input.type = "checkbox"; input.validity.stepMismatch is false
PASS input.type = "color"; input.validity.stepMismatch is false
PASS input.type = "email"; input.validity.stepMismatch is false
PASS input.type = "hidden"; input.validity.stepMismatch is false
PASS input.type = "image"; input.validity.stepMismatch is false
PASS input.type = "khtml_isindex"; input.validity.stepMismatch is false
PASS input.type = "passwd"; input.validity.stepMismatch is false
PASS input.type = "radio"; input.validity.stepMismatch is false
PASS input.type = "reset"; input.validity.stepMismatch is false
PASS input.type = "search"; input.validity.stepMismatch is false
PASS input.type = "submit"; input.validity.stepMismatch is false
PASS input.type = "tel"; input.validity.stepMismatch is false
PASS input.type = "url"; input.validity.stepMismatch is false
PASS input.type = "file"; input.validity.stepMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
